internal/debug: Refactor from `debugLog()` function.
