+++ internal/debug/nodebug.go
@@ -1,0 +1,5 @@
+// +build !p9debug
+
+package debug
+
+func Log(str string, args ...interface{}) {}
