--- nodebug.go
@@ -1,5 +1,0 @@
-// +build !p9debug
-
-package p9
-
-func debugLog(str string, args ...interface{}) {}
